Add monitoring, authentication, menu naming and change file infrastructure
