+++ app/templates/login.html
@@ -1,99 +1,53 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1, shrink-to-fit=no"
-    />
-    <meta property="og:title" content="Flixlist" />
+{% extends "header.html" %}
 
-    <link
-      rel="stylesheet"
-      href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
-      integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T"
-      crossorigin="anonymous"
-    />
-
-    <link
-      rel="stylesheet"
-      href="{{ url_for('static', filename='styles/main.css') }}"
-    />
-
-    <title>{{ title }}</title>
-  </head>
-  <body style="background-color: #070525">
-    <div class="container">
-      <div class="row justify-content-md-center">
-        <div class="col-md-auto">
-          <img
-            src="/static/img/flixlist_logo2.png"
-            alt="..."
-            class="img-fluid"
-          />
-        </div>
-      </div>
-      <div class="row justify-content-md-center">
-        <div class="mx-auto text-center" style="width: 300px">
-          <form class="form-signin" method="POST">
-            <h1
-              class="h3 mb-3 font-weight-normal login-text"
-              style="color: aliceblue"
-            >
-              Please sign in
-            </h1>
-            <label for="inputUser" class="sr-only">User</label>
-            <input
-              type="text"
-              id="inputUser"
-              class="form-control"
-              placeholder="Username"
-              name="username"
-              required
-              autofocus
-            />
-            <label for="inputPassword" class="sr-only">Password</label>
-            <input
-              type="password"
-              id="inputPassword"
-              class="form-control"
-              placeholder="Password"
-              name="password"
-              required
-            />
-            <div class="checkbox mb-3">
-              <label style="color: aliceblue">
-                <input type="checkbox" value="remember-me" />
-                Remember me
-              </label>
-            </div>
-            <button
-              class="btn btn-lg btn-primary btn-dark btn-block"
-              type="submit"
-              style="background-color: #e50914"
-            >
-              Sign in
-            </button>
-            <p class="mt-5 mb-3 text-muted">2021</p>
-          </form>
-        </div>
-      </div>
-    </div>
-    <script
-      src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
-      integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
-      crossorigin="anonymous"
-    ></script>
-    <script
-      src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"
-      integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1"
-      crossorigin="anonymous"
-    ></script>
-    <script
-      src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"
-      integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM"
-      crossorigin="anonymous"
-    ></script>
-  </body>
-</html>
+{% block content %}
+<div class="container">
+	<div class="row justify-content-md-center">
+		<div class="col-md-auto">
+			<img src="/static/img/flixlist_logo2.png" alt="..." class="img-fluid" />
+		</div>
+	</div>
+	<div class="row justify-content-md-center">
+		<div class="mx-auto text-center" style="width: 300px">
+			<form class="form-signin" method="POST">
+				<h1 class="h3 mb-3 font-weight-normal login-text" style="color: aliceblue"> Login </h1>
+				<div class="box">
+					{% with messages = get_flashed_messages() %}
+					{% if messages %}
+					<div class="notification is-danger">
+						{{ messages[0] }}
+					</div>
+					{% endif %}
+					{% endwith %}
+				</div>
+				<label for="inputUser" class="sr-only">User</label>
+				<input type="text" id="inputUser" class="form-control" placeholder="Username" name="username" required
+					autofocus />
+				<label for="inputPassword" class="sr-only">Password</label>
+				<input type="password" id="inputPassword" class="form-control" placeholder="Password" name="password"
+					required />
+				<div class="checkbox mb-3">
+					<label style="color: aliceblue">
+						<input type="checkbox" value="remember-me" />
+						Remember me
+					</label>
+				</div>
+				<button class="btn btn-lg btn-primary btn-dark btn-block" type="submit"
+					style="background-color: #e50914">
+					Sign in
+				</button>
+				<p class="mt-5 mb-3 text-muted">2021</p>
+			</form>
+		</div>
+	</div>
+</div>
+<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
+	integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
+	crossorigin="anonymous"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"
+	integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1"
+	crossorigin="anonymous"></script>
+<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"
+	integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM"
+	crossorigin="anonymous"></script>
+{% endblock %}
