+++ prometheus.yml
@@ -1,0 +1,27 @@
+global:
+  scrape_interval:     15s 
+  evaluation_interval: 15s 
+
+  external_labels:
+      monitor: 'flixlist' 
+
+scrape_configs:
+  - job_name: prometheus
+    scheme: https
+    metrics_path: /prometheus
+    static_configs:
+      - targets: ['flixlist.tech']
+
+  - job_name: node-exporter
+    scheme: https
+    metrics_path: /node-exporter
+    static_configs:
+      - targets: ['flixlist.tech']
+
+  - job_name: cadvisor
+    scrape_interval: 5s
+    scheme: https
+    metrics_path: /cadvisor
+    static_configs:
+      - targets: ['flixlist.tech']
+
