+++ user_conf.d/nginx-conf.conf
@@ -2,10 +2,39 @@ server {
     listen 443 ssl;
     listen [::]:443 ssl;
 
-    server_name flixlist.tech www.flixlist.tech;
+    server_name flixlist.tech www.flixlist.tech flixlist.duckdns.org;
+
+    location ~ ^/cadvisor(.*)$ {
+        proxy_pass http://cadvisor:8080$1;
+        proxy_redirect / /cadvisor/;
+        resolver 127.0.0.11;
+    }
+
+    location ~ ^/prometheus(.*)$ {
+	proxy_pass http://prometheus:9090$1;
+	proxy_redirect / /prometheus/;
+	resolver 127.0.0.11;
+    }
+
+    location ~ ^/node-exporter(.*)$ {
+        add_header Content-Security-Policy "script-src: 'unsafe-eval' 'unsafe-inline';";
+	proxy_pass http://node-exporter:9100$1;
+	resolver 127.0.0.11;
+    }
+
+    location ~ ^/grafana(.*)$ {
+        add_header Content-Security-Policy "script-src: 'unsafe-eval' 'unsafe-inline';";
+	proxy_pass http://grafana:3000$1;
+	resolver 127.0.0.11;
+    }
 
     location / {
         proxy_pass http://web:80/;
+        proxy_set_header Host $http_host;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-Proto $scheme;
+        proxy_redirect http://$http_host/ https://$http_host/;
     }
 
     # Load the certificate files.
@@ -13,13 +42,11 @@ server {
     ssl_certificate_key /etc/letsencrypt/live/flixlist/privkey.pem;
     ssl_trusted_certificate /etc/letsencrypt/live/flixlist/chain.pem;
 
-    # Load the Diffie-Hellman parameter.
-    ssl_dhparam /etc/letsencrypt/dhparams/dhparam.pem;
 }
 
 server {
     listen 80;
-    server_name flixlist.tech www.flixlist.tech;
+    server_name flixlist.tech www.flixlist.tech flixlist.duckdns.org;
 
     location ~ /.well-known/acme-challenge {
         allow all;
@@ -31,6 +58,10 @@ server {
     }
 
     if ($host = flixlist.tech) {
+        return 301 https://$host$request_uri;
+    }
+
+    if ($host = flixlist.duckdns.org) {
         return 301 https://$host$request_uri;
     }
 }
